pyproject.toml
[project]
name = "ktp-conductor-demo"
version = "0.1.0"
description = "Aerial conductor segmentation, vectorisation, hidden-cable inference, and LiDAR classification — research prototype"
readme = "README.md"
requires-python = ">=3.11,<3.13"
license = { text = "MIT" }
dependencies = [
    # Web stack
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "python-multipart>=0.0.12",
    "pydantic>=2.9",
    "pydantic-settings>=2.5",
    "slowapi>=0.1.9",
    "prometheus-fastapi-instrumentator>=7.0",
    "structlog>=24.4",
    # ML / CV
    "torch>=2.4",
    "torchvision>=0.19",
    "segmentation-models-pytorch>=0.3.4",
    "albumentations>=1.4",
    "opencv-python-headless>=4.10",
    "scikit-image>=0.24",
    "pillow>=10.4",
    "numpy>=1.26,<2.1",
    "scipy>=1.14",
    # Geo
    "shapely>=2.0",
    "networkx>=3.3",
    "rasterio>=1.3",
    "geopandas>=1.0",
    # LiDAR
    "laspy>=2.5",
    "lazrs>=0.6",
    # Utilities
    "tqdm>=4.66",
    "loguru>=0.7",
    "rich>=13.8",
]

[project.urls]
Repository = "https://github.com/s1ddh-rth/ktp-conductor-demo"
Issues = "https://github.com/s1ddh-rth/ktp-conductor-demo/issues"

[project.optional-dependencies]
training = [
    "pytorch-lightning>=2.4",
    "torchmetrics>=1.4",
    "wandb>=0.18",
    "kaggle>=1.6",
]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "ruff>=0.6",
    "ipython>=8.27",
    "jupyter>=1.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "A", "C4", "SIM"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
